Reconfigured auth routes
