+++ routes/authRoutes.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { signup, login } from "../controllers/authController.js";
+import { signup, signin } from "../controllers/authController.js";
 
 const router = express.Router();
 
@@ -7,6 +7,6 @@ const router = express.Router();
 router.post("/signup", signup);
 
 // Route for user login
-router.post("/login", login);
+router.post("/login", signin);
 
 export default router;
